CogBench/requirements.txt
#numba==0.59.0
numba==0.60.0
numpy==1.26.4
#openai==0.27.4
#pandas==2.2.0
pandas==2.2.1
python-dotenv==1.0.1
#statsmodels==0.14.1
statsmodels==0.14.3
#anthropic==0.19.2
torch==2.1.2
gymnasium==0.26.3 #Or gymnasium==0.19.0
tqdm==4.66.2 
#google-cloud-aiplatform==1.42.1
transformers==4.37.2
#umap-learn==0.5.5
umap-learn==0.5.6
seaborn==0.13.2